requirements.txt
Pillow>=10.0.0
hitherdither @ git+https://github.com/hbldh/hitherdither
